refactor(hero): co-locate component, CSS, unit test, E2E in Hero/ folder
