+++ components/sections/Hero/index.ts
@@ -1,0 +1,1 @@
+export { Hero } from './Hero';
